# Conflicts:
#	.idea/dataSources.xml
#	gomen/src/main/java/com/tuna/gomen/board/repository/BoardRepository.java
#	gomen/src/main/java/com/tuna/gomen/user/repository/UserRepository.java

+++ gomen/src/main/java/com/tuna/gomen/mapper/BookmarkMapper.java
@@ -7,6 +7,6 @@ import java.util.List;
 
     @Mapper
     public interface BookmarkMapper {
-        List<BookmarkedBoardDto> findBookmarkedPostsByUserId(@Param("userId") Long userId);
+        List<BookmarkedBoardDto> findBookmarkedPostsByUserId(@Param("userId") Integer userId);
     }
